progress on multi-slot mode
