+++ src/BpmMain.cpp
@@ -1,5 +1,6 @@
 #include <Arduino.h>
 #include <Wire.h>
+#include <iostream>
 
 // #include "./include/bpmWiFi.hpp"
 #include "./include/bpmSensor.hpp"
@@ -7,49 +8,37 @@
 
 // max https://datasheets.maximintegrated.com/en/ds/MAX86150.pdf
 
-// hf::BpmSensor bpmSensor = hf::BpmSensor(1);
-// // hf::BpmWiFi bpmWiFi;
+hf::BpmSensor bpmSensor = hf::BpmSensor(3);
+unsigned long curTime;
 
-// void setup()
-// {   
-//     Serial.begin(115200);
-//     Wire.begin();
-//     Wire.setClock(400000);
-//     delay(10);
+// hf::BpmWiFi bpmWiFi;
 
-//     // bpmWiFi.initWiFi();
-//     bpmSensor.init();
-
-//     // loop until serial connection opens - diagnostic
-//     while (!Serial)
-//         delay(1);
-    
-//     Serial.println("Setup Complete");
-//     delay(100);
-// }
-
-// void loop()
-// {
-//     bpmSensor.sample();
-//     // delay(1000);
-// }
-
-hf::StressTest test = hf::StressTest();
-
-void setup() 
-{
+void setup()
+{   
     Serial.begin(115200);
     Wire.begin();
     Wire.setClock(400000);
+    delay(10);
 
-    while (!Serial) delay(1);
+    // bpmWiFi.initWiFi();
+    bpmSensor.init();
+
+    // loop until serial connection opens - diagnostic
+    while (!Serial)
+        delay(1);
+    
+    Serial.println("Setup Complete");
+    delay(100);
+
+    curTime = millis();
 }
 
 void loop()
 {
-    Serial.println("BASIC TEST STARTING");
-    test.basic(); // basic calculatoins - reads and sends data
-
-    Serial.println("INTENSE TEST STARTING");
-    test.intense();
+    bpmSensor.sample();
+    // if (millis() > curTime + 1000) {
+    //     bpmSensor.sample();
+    //     curTime += 1000;
+    // }
+    // delay(1000);
 }
