+++ src/include/constants.hpp
@@ -1,2 +1,11 @@
 #define I2C_ADDRESS 94
 #define WINDOW_LENGTH 256
+
+#define BUFFER_LENGTH 32
+
+// PPG & ECG SLOTS
+// CUR CODE IS BASED OFF THIS SLOT STRUCTURE
+// slot 0 and slot 1 can be either red or ir depending on config
+#define PPG_SLOT0 1
+#define ECG_SLOT  2
+#define PPG_SLOT1 3
